TP_SD/requirements.txt
django-adminlte3
djangorestframework
markdown
django-filter
numpy
pandas